Worked on the unfolding plot with actions. Resulting in refactoring the package to make modules clearer.
